refactor(core): fix BaseQuery and BaseService abstraction
