+++ app/core/queries/base.py
@@ -1,0 +1,44 @@
+from sqlmodel import Session, select
+from typing import Type, Generic, Optional, List
+
+from ..types import ModelType
+
+
+class BaseQuery(Generic[ModelType]):
+    def __init__(self, db: Session, model: Type[ModelType]):
+        self.db = db
+        self.model = model
+
+    def get_all(self) -> List[ModelType]:
+        statement = select(self.model)
+        return self.db.exec(statement).all()
+
+    def get_by_id(self, id: int) -> Optional[ModelType]:
+        return self.db.get(self.model, id)
+
+
+class CreateQueryMixin(Generic[ModelType]):
+    def create(self, data: ModelType) -> ModelType:
+        self.db.add(data)
+        self.db.commit()
+        self.db.refresh(data)
+
+        return data
+
+
+class UpdateQueryMixin(Generic[ModelType]):
+    def update(self, record: ModelType, data: dict) -> ModelType:
+        record.sqlmodel_update(data)
+        self.db.add(record)
+        self.db.commit()
+        self.db.refresh(record)
+
+        return record
+
+
+class DeleteQueryMixin(Generic[ModelType]):
+    def delete(self, record: ModelType) -> ModelType:
+        self.db.delete(record)
+        self.db.commit()
+
+        return record
