+++ app/core/types.py
@@ -1,0 +1,5 @@
+from sqlmodel import SQLModel
+from typing import TypeVar
+
+ModelType = TypeVar("ModelType", bound=SQLModel)
+CreateModelType = TypeVar("CreateModelType", bound=SQLModel)
