+++ app/features/users/queries.py
@@ -1,12 +1,20 @@
 from sqlmodel import Session
 
-from app.core.models import BaseQuery
+from app.core.queries.base import (
+    BaseQuery,
+    CreateQueryMixin,
+    UpdateQueryMixin,
+    DeleteQueryMixin,
+)
 
 from .models import User
-from .schemas import UserCreate, UserUpdate
 
 
-class UserQueries(BaseQuery[User, UserCreate, UserUpdate]):
+class UserQueries(
+    BaseQuery[User],
+    CreateQueryMixin[User],
+    UpdateQueryMixin[User],
+    DeleteQueryMixin[User],
+):
     def __init__(self, db: Session):
         super().__init__(db, User)
-
