+++ app/features/users/services.py
@@ -1,7 +1,7 @@
 from fastapi import status
 
 from app.core.constants import CRUDMessages
-from app.core.models import BaseService
+from app.core.services.base import BaseService, CreateServiceMixin
 from app.core.schemas.http import HTTPResponseModel
 
 from .models import User
@@ -9,13 +9,44 @@ from .queries import UserQueries
 from .schemas import UserCreate, UserUpdate
 
 
-class UserService(BaseService[User, UserCreate, UserUpdate]):
+class UserService(BaseService[User], CreateServiceMixin[UserCreate]):
     queries = UserQueries
 
     def __init__(self, queries):
         super().__init__(queries, User)
 
+    def update(self, id: int, data: UserUpdate):
+        record = self.queries.get_by_id(id)
 
+        if record:
+            new_data = data.model_dump(exclude_unset=True)
+            updated = self.queries.update(record, new_data)
+            return HTTPResponseModel(
+                status_code=status.HTTP_200_OK,
+                message=CRUDMessages.UPDATE_SUCCESS,
+                data=updated,
+            )
 
+        return HTTPResponseModel(
+            status_code=status.HTTP_404_NOT_FOUND,
+            message=CRUDMessages.UPDATE_FAILED,
+            errors=[{"detail": CRUDMessages.GET_NOT_FOUND}],
+        )
 
+    def delete(self, id: int):
+        record = self.queries.get_by_id(id)
 
+        if record:
+            deleted = self.queries.delete(record)
+
+            return HTTPResponseModel(
+                status_code=status.HTTP_200_OK,
+                message=CRUDMessages.DELETE_SUCCESS,
+                data=deleted,
+            )
+
+        return HTTPResponseModel(
+            status_code=status.HTTP_404_NOT_FOUND,
+            message=CRUDMessages.DELETE_FAILED,
+            errors=[{"detail": CRUDMessages.GET_NOT_FOUND}],
+        )
